Adam's attempt at a start
I wasn't able to get the API working due to a CORS error that I can't fix. Either way, I started...something. I put in Vuejs (which won't really work until there is a functioning API attached) and put in a search button and table to use as a display. Neither are really populated yet.