LoL API gets recent match data
LoL API working and implemented. However, data is not generally understandable (champs are ints, etc). Additionally, an error is thrown when an incorrect summoner name is used (such as asdfgjkadfshnj or NotARealSummonerName).
Match data is set in its own function because a different LoL API needs to be used by getting account number.